Updated Overlay
You can now click on the various HUD icons and the console will output
the model used for that icon (try scrolling and clicking different
shapes).  This can be used for proper menu buttons.  I will work on
cleaning up the code and implementing those buttons.

Also, I added a to-do list of 2D assets for Gerald to do.