Basic Button
Okay, I found a function that actually does buttons, so we don't have to
make them ourselves.  Not sure how to go about running the menu
"outside" of the game, though.  Will figure that out tomorrow during
class (or maybe beforehand).

Apologies for not getting much done for the past few days, been fairly
busy with other projects and homework.