Finished Menu and Overlay (Mostly)
Okay, both are almost done.  Just need to do Options once we actually
have options to change (or we can remove that button), and fix up the
HUD sprites.

I am also still having a pathing error with the new environment, which
is REALLY weird considering the exact same path works with the old
one...  But anyway, please leave the "zzzzzzarena" thing there for now.